requirements.txt
streamlit
requests           # Para acceder a la API
plotly             # Para gráficos interactivos
pandas             # Para el procesamiento de datos
python-dotenv	   # Para la api key